Move `Gear` to the `character` package
