+++ src/main/kotlin/profile/ProfileGearSet.kt
@@ -1,7 +1,5 @@
 package cloud.drakon.ktlodestone.profile
 
-import cloud.drakon.ktlodestone.profile.gear.Gear
-
 data class ProfileGearSet(
     val mainHand: Gear,
     val offHand: Gear?,
@@ -14,5 +12,5 @@ data class ProfileGearSet(
     val necklace: Gear?,
     val bracelets: Gear?,
     val ring: List<Gear?>?,
-    val soulCrystal: String?
+    val soulCrystal: String?,
 )
